% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCCP_single_changepoint.R
\name{DCCP_single_changepoint}
\alias{DCCP_single_changepoint}
\title{Detect a Single Change Point}
\usage{
DCCP_single_changepoint(
  data,
  FUN_single_changepoint = M.L1norm,
  nperm = 200,
  sig.lvl = 0.05
)
}
\arguments{
\item{data}{A numeric matrix or data frame with observations in rows and
variables in columns.}

\item{FUN_single_changepoint}{A function used to calculate pairwise
distances. To use the modified L2 distance, set
`FUN_single_changepoint = dist`.}

\item{nperm}{Number of permutations.}

\item{sig.lvl}{Significance level.}
}
\value{
A list containing the detected change point and permutation p-value.
}
\description{
Estimates a single change point and assesses its significance using a
permutation test.
}
