% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{MITcellphone}
\alias{MITcellphone}
\title{MIT Cellphone Activity Data}
\format{
A numeric matrix with 232 rows and 4,560 columns.
}
\source{
MIT Reality Mining project.
}
\description{
Daily pairwise cellphone activities for 96 participants from 15 September
2004 to 4 May 2005. Each of the 4,560 columns represents one unique pair of
participants and each row represents one of 232 days.
}
\references{
Eagle, N. and Pentland, A. (2006). Reality mining: sensing
  complex social systems. Personal and Ubiquitous Computing, 10, 255-268.
}
