% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SP500data}
\alias{SP500data}
\title{Standardized S&P500 Stock Prices}
\format{
A numeric matrix with 108 rows and 496 columns.
}
\source{
Yahoo Finance, obtained using the \code{BatchGetSymbols} package.
}
\description{
Standardized daily closing prices for 496 S&P500 companies over 108 trading
days from 1 January 2020 to 29 May 2020. The same source data are also used
and available in the \code{HDDchangepoint} package.
}
\references{
Zhang, L. and Drikvandi, R. (2025). Distance-based CUSUM
  statistics for high dimensional change points. Statistics and Computing,
  35, Article 215. \doi{10.1007/s11222-025-10752-1}.
}
